Moved some fonctions to sim_lib, and changed the caluclate_overall_average code.
